+++ components/places/src/db/db.rs
@@ -43,15 +43,20 @@ impl PlacesDb {
         // which the SQLcipher docs themselves say is too small and should be changed.
         let encryption_pragmas = if let Some(key) = encryption_key {
             format!(
-                "
-                PRAGMA key = '{key}';
-                PRAGMA cipher_page_size = {page_size};
-            ",
+                "PRAGMA key = '{key}';
+                 PRAGMA cipher_page_size = {page_size};",
                 key = sql_support::escape_string_for_pragma(key),
                 page_size = PAGE_SIZE,
             )
         } else {
-            format!("PRAGMA page_size = {};", PAGE_SIZE)
+            format!(
+                "PRAGMA page_size = {};
+                 -- Disable calling mlock/munlock for every malloc/free.
+                 -- In practice this results in a massive speedup, especially
+                 -- for insert-heavy workloads.
+                 PRAGMA cipher_memory_security = false;",
+                PAGE_SIZE
+            )
         };
 
         let initial_pragmas = format!(
@@ -62,6 +67,8 @@ impl PlacesDb {
             -- files in memory, since on Android there's no tmp partition. See
             -- https://github.com/mozilla/mentat/issues/505. Ideally we'd only
             -- do this on Android, and/or allow caller to configure it.
+            -- (although see also bug 1313021, where Firefox enabled it for both
+            -- Android and 64bit desktop builds)
             PRAGMA temp_store = 2;
 
             -- 6MiB, same as the value used for `promiseLargeCacheDBConnection` in PlacesUtils,
@@ -72,6 +79,9 @@ impl PlacesDb {
 
             -- We want foreign-key support.
             PRAGMA foreign_keys = ON;
+
+            -- we unconditionally want write-ahead-logging mode
+            PRAGMA journal_mode=WAL;
         ",
             encryption_pragmas,
         );